requirements.txt
dbus-python>=1.3.2
click>=8.1.0
rich>=13.0.0
textual>=0.34.0
pyyaml>=6.0